Add utilities_test.py
